+++ witp/src/main/java/dev/efnilite/ip/schematic/Schematics.java
@@ -2,17 +2,10 @@ package dev.efnilite.ip.schematic;
 
 import dev.efnilite.ip.IP;
 import dev.efnilite.ip.schematic.legacy.LegacySchematicMigrator;
-import dev.efnilite.ip.util.Colls;
 import dev.efnilite.vilib.util.Task;
 import dev.efnilite.vilib.util.Time;
 
 import java.io.File;
-import java.io.InputStream;
-import java.net.URL;
-import java.net.UnknownHostException;
-import java.nio.file.FileAlreadyExistsException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.*;
 
 /**
@@ -20,9 +13,18 @@ import java.util.*;
  */
 public class Schematics {
 
-    private static final int SCHEMATIC_COUNT = 21;
-
-    private static final String URL = "https://github.com/Efnilite/Walk-in-the-Park/raw/main/schematics/";
+    private static final String[] SPAWN_SCHEMATICS = new String[] {
+            "spawn-island", "spawn-island-duels"
+    };
+    private static final String[] PARKOUR_SCHEMATICS = new String[] {
+            "686ebc21", "428fcf54", "323ab385", "b79cc16b", "31df2b96",
+            "f9d2c0ac", "d166d78f", "0507171d", "93570681", "bd89f053",
+            "98d1f8e8", "e923ac20", "74b4fd1e", "befcc649", "1d1a7f00",
+            "e037702e", "e772a8f0", "2bd47c9f", "432623a9", "34037304",
+            "bd7b4c19", "6fdae317", "51f1e203", "6797dbf9", "135f8b2b",
+            "365409ca", "72ec530f", "8a7575ec", "fa035723", "1d5969be",
+            "09ba8fa3", "1dc87b41", "bb99d8bc", "ea9f7e05", "8dfd98c1"
+    };
 
     private static final File FOLDER = IP.getInFolder("schematics");
 
@@ -62,26 +64,10 @@ public class Schematics {
         }).run();
     }
 
-    // downloads all schematics
-    public static void download() {
-        IP.logging().info("Downloading schematics...");
-
+    private static void download() {
         List<String> schematics = new ArrayList<>();
-        schematics.addAll(Arrays.asList("spawn-island.witp", "spawn-island-duels.witp"));
-        schematics.addAll(Colls.range(1, SCHEMATIC_COUNT + 1).stream().map("parkour-%d.witp"::formatted).toList());
-        schematics.forEach(Schematics::downloadFile);
-    }
-
-    // downloads a singular file to the schematics folder
-    private static void downloadFile(String name) {
-        try (InputStream stream = new URL("%s%s".formatted(URL, name)).openStream()) {
-            Files.copy(stream, Paths.get(FOLDER.toString(), name));
-        } catch (FileAlreadyExistsException ex) {
-            // do nothing
-        } catch (UnknownHostException ex) {
-            IP.logging().stack("Internet error while downloading schematic %s".formatted(name), ex);
-        } catch (Exception ex) {
-            IP.logging().stack("Error while downloading schematic %s".formatted(name), ex);
-        }
+        schematics.addAll(Arrays.asList(SPAWN_SCHEMATICS));
+        schematics.addAll(Arrays.stream(PARKOUR_SCHEMATICS).map("parkour-%s"::formatted).toList());
+        schematics.forEach(file -> IP.getPlugin().saveResource("schematics/%s".formatted(file), true));
     }
 }